+++ ChangeLog.md
@@ -34,7 +34,9 @@
     - Replace 'addtionalTabs, 'tabsUpdate', 'tabsBefore', 'tabs', 'haveIdent' and 'haveRoles' amAccount parameters by 'editTabsFn'
     - Remove unused AccountsManager.s.getBy() function (and redondant regarding AccountsCore.s.byQuery())
     - Exported AccountsManager.amAccount class is renamed to AccountsManager.Account
-    - Introduce editAdminNotes ReactiveVar (todo #3)"
+    - Introduce editAdminNotes ReactiveVar (todo #3)
+    - Delegating onCreateUser() to AccountsCore means removing onCreateUser() from AccountsManager
+    - Remove methods, only relying on AccountsCore for server accesses
 
 ### 2.2.0